third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-overridden-2-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test for crbug.com/274059. Compare rules from a style in a shadow tree with ::content in a different shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(getNodeInComposedTree("hostChild")).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
